backend/requirements.txt
Django==5.2.7
djangorestframework==3.15.1
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
dj-database-url==2.1.0
psycopg2-binary==2.9.9